Fix the DNS server name issue for OpenStack
With the Kuryr patch, we have decoupled the Heat stack name from the
two Ansible variables that controlled the cluster's domain.

However, the stack name was used in the Heat templates to name the Nova
VMs and this mismatch caused an issue when an external DNS was being
used.

This replaces uses of `openshift_openstack_stack_name` with
`openshift_openstack_full_dns_domain` in the Heat templates.

Fixes #7404